Refactor Code for getBalances
